server/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Telegram (using aiogram - faster and more stable)
aiogram>=3.3.0

# QR Code
qrcode[pil]>=7.4
pyzbar>=0.1.9
Pillow>=10.0.0

# Configuration
python-dotenv>=1.0.0
